mod TestCoroutineInterceptor to use runTest on K/N
