+++ kotest-framework/kotest-framework-engine/src/desktopMain/kotlin/io/kotest/engine/test/interceptors/TestDispatcherInterceptor.kt
@@ -11,11 +11,3 @@ actual class TestDispatcherInterceptor : TestExecutionInterceptor {
       test: suspend (TestCase, TestScope) -> TestResult
    ): TestResult = test(testCase, scope)
 }
-
-actual class TestCoroutineInterceptor : TestExecutionInterceptor {
-   override suspend fun intercept(
-      testCase: TestCase,
-      scope: TestScope,
-      test: suspend (TestCase, TestScope) -> TestResult
-   ): TestResult = test(testCase, scope)
-}
